+++ src/course-manipulation/NewCourseDialog.jsx
@@ -36,9 +36,9 @@ import {
     Typography,
 } from "@mui/material";
 
-import Assessment from "./Assessment";
+import Assessment from "../classes/Assessment";
 import Axios from "axios";
-import ConfirmDialog from "./ConfirmDialog";
+import ConfirmDialog from "../ConfirmDialog";
 import CreateAssessmentCard from "./CreateAssessmentCard";
 import { isMobile } from "react-device-detect";
 import { TransitionGroup } from "react-transition-group";
