+++ src/course-manipulation/SyncAssessmentCard.jsx
