+++ src/course-viewer/CourseViewerEditorDesktop.jsx
