+++ src/course-viewer/CourseViewerEditorMobile.jsx
