+++ src/course-viewer/CourseViewerFilterMobile.jsx
