+++ src/course-viewer/CourseViewerHeader.jsx
