+++ src/course-viewer/CourseViewerMobileActionButtons.jsx
